Remove docker-compose from python dependencies
As this module is unmaintained and is creating conflicts with other
packages, we need to remove it in order to still be able to use
the rest of the collection. In a follow-up we should migrate our
docker_compose module to use docker compose cli, so it would work
with v2 of it.