Don't emit waiting events for healthy containers
The output of the compose command that is used for parsing events can contain multiple events for containers with health checks.  With one event status Waiting then a following event with the status Healthy. This causes the module to erroneously report changes.

This change removes waiting events for healthy containers